2/ml_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9
# Optional ML backends (install one):
# tensorflow==2.15.0
# torch==2.4.0
# torchvision==0.19.0
pytest==8.3.3
httpx==0.27.2
